megred with master

+++ src/Tasks_ExampleMessageTask.hpp
@@ -52,7 +52,7 @@ private:
 
    void receivedCb(Facilities::MeshNetwork::NodeId nodeId, String& msg);
 
-    void draw_img(int img[32][32], int intensity);
+    void draw_img(int img[32][32], uint8_t intensity);
 
 };
